requirements.txt
telethon
python-dotenv
gspread
oauth2client
requests
asyncio
